Relocated utils.go into exchange, and made the functions private.
